fix login switch
